requirements.txt
python-dotenv==0.18.0
python-binance==1.0.12
tradingview-ta==3.2.5